refactor QueryFilterModel imports to use separated types file
