+++ ui/apps/vue-mri-ui-lib/src/query-filter/components/QueryFilterEvent.vue
@@ -6,7 +6,7 @@ export default {
 
 <script setup lang="ts">
 import { defineProps, defineEmits } from 'vue'
-import { QueryFilterEvent } from '../models/QueryFilterModel'
+import { QueryFilterEvent } from '../types/QueryFilterTypes'
 import AttributesDropdown from './AttributesDropdown.vue'
 import { type AttributeOption } from '../utils/CriteriaConfigLoader'
